@@ -18,7 +18,7 @@
     "coverage:native": "node scripts/test.js --config \"./config/native.jest.config.js\" --coverage"
   },
   "dependencies": {
-    "react": "^16.6.x",
+    "react": "^16.9.0",
     "react-app-polyfill": "^0.1.3",
     "react-art": "^16.6.x",
     "react-dom": "^16.6.x",
