@@ -20,6 +20,11 @@ const styles = StyleSheet
     },
   );
 
+const sanitize = (str = '') => str
+  .replace(/#/g, ' #')
+  .replace(/@/g, ' @')
+  .replace(/\s\s+/g, ' ');
+
 const TwitterTextView = ({
   children = '',
   extractHashtags,
@@ -33,7 +38,7 @@ const TwitterTextView = ({
   <Text
     {...extraProps}
   >
-    {children
+    {sanitize(children)
       .split(' ')
       .map(
         (word, i) => {
@@ -43,12 +48,21 @@ const TwitterTextView = ({
               word,
             ) || [];
             if (hashtag) {
+              const result = `${pfx}#${hashtag}`;
+              const after = `${word.substring(hashtag.length + 1)}`;
               return (
                 <Text
-                  onPress={e => onPressHashtag(e, hashtag)}
-                  style={hashtagStyle}
                 >
-                  {`${pfx}#${hashtag}`}
+                  <Text
+                    onPress={e => onPressHashtag(e, hashtag)}
+                    style={hashtagStyle}
+                  >
+                    {`${result}`}
+                  </Text>
+                  <Text
+                  >
+                    {`${after}`}
+                  </Text>
                 </Text>
               );
             }
@@ -58,12 +72,21 @@ const TwitterTextView = ({
               word,
             ) || [];
             if (mention) {
+              const result = `${pfx}@${mention}`;
+              const after = `${word.substring(mention.length + 1)}`;
               return (
                 <Text
-                  onPress={e => onPressMention(e, mention)}
-                  style={mentionStyle}
                 >
-                  {`${pfx}@${mention}`}
+                  <Text
+                    onPress={e => onPressMention(e, mention)}
+                    style={mentionStyle}
+                  >
+                    {`${result}`}
+                  </Text>
+                  <Text
+                  >
+                    {`${after}`}
+                  </Text>
                 </Text>
               );
             }
@@ -93,19 +116,21 @@ TwitterTextView.defaultProps = {
   children: '',
   extractHashtags: true,
   onPressHashtag: (e, hashtag) => {
+    const msg = `Hashtag: "${hashtag}"`;
     if (Platform.OS !== 'web') {
-      Alert.alert(hashtag);
+      Alert.alert(msg);
     } else {
-      console.log(hashtag);
+      console.log(msg);
     }
   },
   hashtagStyle: styles.linkStyle,
   extractMentions: true,
   onPressMention: (e, mention) => {
+    const msg = `Mention: "${mention}"`;
     if (Platform.OS !== 'web') {
-      Alert.alert(mention);
+      Alert.alert(msg);
     } else {
-      console.log(mention);
+      console.log(msg);
     }
   },
   mentionStyle: styles.linkStyle,
